Moved Types to /src/types
