+++ frontend/src/components/Confirm/Confirm.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 // eslint-disable-next-line no-unused-vars
-import { ConfirmProps } from '../../lib/types/ConfirmProps';
+import { ConfirmProps } from '../../types/ConfirmProps';
 import { ModalWindow } from '../ModalWindow/ModalWindow';
 
 import './Confirm.css';
