+++ frontend/src/components/EditDietSettingsForm/EditDietSettingsForm.tsx
@@ -13,7 +13,7 @@ import { AppContext } from '../AppContext/AppContext';
 import { LoadingBox } from '../LoadingBox/LoadingBox';
 import { MacroTargetForm } from '../MacroTargetForm/MacroTargetForm';
 // eslint-disable-next-line no-unused-vars
-import { DietSettingsUpdateRequest } from '../../lib/types/dietSettingsUpdateRequest';
+import { DietSettingsUpdateRequest } from '../../types/DietSettingsUpdateRequest';
 
 const EditDietSettingsForm = () => {
     const [status, setStatus] = useState('initialized');
