+++ frontend/src/components/GearGrid/GearGrid.test.tsx
@@ -3,7 +3,7 @@ import { render, screen } from '@testing-library/react';
 import { AppContext } from '../AppContext/AppContext';
 import { GearGrid } from './GearGrid';
 // eslint-disable-next-line no-unused-vars
-import { Gear } from '../../lib/types/Gear';
+import { Gear } from '../../types/Gear';
 
 describe('<GearGrid />', () => {
     test('it renders the grid', async () => {
