+++ frontend/src/components/Grid/Grid.tsx
@@ -1,9 +1,9 @@
 import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 // eslint-disable-next-line no-unused-vars
-import { GridProps } from '../../lib/types/GridProps';
+import { GridProps } from '../../types/GridProps';
 // eslint-disable-next-line no-unused-vars
-import { GridColumn } from '../../lib/types/GridColumn';
+import { GridColumn } from '../../types/GridColumn';
 
 import './Grid.css';
