+++ frontend/src/components/TextBox/TextBox.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { FormValidator } from '../../lib/FormValidator';
 // eslint-disable-next-line no-unused-vars
-import { TextBoxProps } from '../../lib/types/TextBoxProps';
+import { TextBoxProps } from '../../types/TextBoxProps';
 
 const TextBox = (props: TextBoxProps) => {
     const {
