+++ frontend/src/types/ActivityDataRow.ts
