+++ frontend/src/types/DietSettingsUpdateRequest.ts
