+++ frontend/src/types/Gear.ts
