+++ frontend/src/types/GearDataRow.ts
