+++ frontend/src/types/GridColumn.ts
