+++ frontend/src/types/GridProps.ts
