+++ frontend/src/types/GridRow.ts
