+++ frontend/src/types/TextBoxProps.ts
