+++ frontend/src/types/ValidationResult.ts
